Style: Reorganizacion de archivos de app y core (strategy, order y service
