+++ .gitignore
@@ -53,3 +53,6 @@ dkms.conf
 
 # debug information files
 *.dwo
+
+# visual studio
+*.vs
