+++ DeliveryGo/DeliveryGo/Application/Order/Observers/AuditObserver.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DeliveryGo.Core.Order.Observers
+namespace DeliveryGo.Application.Order.Observers
 {
-    public class LogisticsObserver
+    public class AuditObserver
     {
     }
 }
