+++ DeliveryGo/DeliveryGo/Application/Services/Implementations/OrderService.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DeliveryGo.Application.Services.Implementations
+{
+    public class OrderService
+    {
+
+    }
+}
