--- DeliveryGo/DeliveryGo/Core/Services/Implementations/ShippingService.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DeliveryGo.Core.Services.Implementations
-{
-    public class ShippingService
-    {
-    }
-}
